+++ Gemfile
@@ -9,6 +9,7 @@ gemspec
 # Remove verson constraint once latter versions release their -java packages
 gem "bootsnap", "= 1.4.9"
 gem "dry-monitor"
+gem "dry-types"
 gem "zeitwerk"
 
 group :tools do
