+++ lib/dry/system/provider/source_dsl.rb
@@ -33,6 +33,20 @@ module Dry
         def stop(&block)
           source_class.define_method(:stop, &block)
         end
+
+        private
+
+        def method_missing(name, *args, &block)
+          if source_class.respond_to?(name)
+            source_class.public_send(name, *args, &block)
+          else
+            super
+          end
+        end
+
+        def respond_to_missing?(name, include_all = false)
+          source_class.respond_to?(name, include_all) || super
+        end
       end
     end
   end
